apps/api/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
pandas==2.2.3
httpx==0.27.2
python-dotenv==1.0.1
posthog==3.8.0
openai==1.59.7
plotly==5.24.1
pytest==8.3.4
pytest-asyncio==0.24.0
